(feat):server seed/model
