+++ script/seed.js
@@ -149,12 +149,12 @@ var data = {
   ],
   wine: [
 
-{name: 'Caymus', vintage: 2009, quantity: 2, description: 'Supple and creamy-textured, with mocha, ripe plum and cherry flavors that show touches of tobacco, espresso and underbrush. Full-bodied, gaining tannic traction on the finish, elegant and readily approachable.', varietal: 'Cabernet Sauvignon', price: 115.00, size: '750 ml', img: '/images/caymus09.jpeg', place: { city: 'Napa Valley', state: 'California', country: 'USA'}},
+{name: 'Caymus', vintage: 2009, quantity: 2, description: 'Supple and creamy-textured, with mocha, ripe plum and cherry flavors that show touches of tobacco, espresso and underbrush. Full-bodied, gaining tannic traction on the finish, elegant and readily approachable.', varietal: 'Cabernet Sauvignon', price: '115.00', size: '750 ml', img: '/images/caymus09.jpeg', place: { city: 'Napa Valley', state: 'California', country: 'USA'}},
 {name: 'Flowers', vintage: 2014, quantity: 2, description: 'Deep garnet hue. Aromas of wild strawberry, pomegranate, cranberry, savory dried herbs, violet and forest floor lift elegantly from the glass. On the palate, layers of blood orange and hibiscus are framed by fine-grained tannins, coastal minerality and brilliant acidity that culminate in a lengthy finish.', varietal: 'Pinot Noir', price: '90.00', size: '750 ml', img: '/images/flowerspinot.png', place: { city: 'Sonoma Coast', state: 'California', country: 'USA'}},
 {name: 'Kistler', vintage: 2013, quantity: 2, description: 'Deep, concentrated, full-bodied notes of white currants, lemon oil, apple blossom, white peach and honeysuckle are all present in this beautiful wine that will drink nicely for the next seven or so years.', varietal: 'Chardonnay', price: '149.99', size: '750 ml', img: '/images/kistler.jpg', place: { city: 'Sonoma Coast', state: 'California', country: 'USA'}},
-{name: 'Stags Leap', vintage: 2012, quantity: 10, description: 'This Napa Valley Cabernet Sauvignon exemplifies the elegant success that resulted from the cooler 2011 vintage. Black currant, cherry and blueberry aromatics remain distinctly bright in this inviting wine. The palate offers a layered complexity of forest floor, savory herbs, and sweet tobacco that seamlessly blends with the vibrant red fruit. Persistent tannins frame a round, full structure and soft, fresh finish.', varietal: 'Cabernet Saugvinon', price: '107.99', size: '750 ml', img: '/images/stags.png', place:{ city: 'Napa Valley', state: 'California', country: 'USA'}},
-{name: 'Screaming Eagle', vintage: 2012, quantity: 4, description: 'Bright medium ruby. Very deep aromas of blueberry, flowers, licorice and wild herbs show a medicinal reserve. Thick but vibrant, with deep, fully ripe, layered flavors of blackberry and blueberry syrup and graphite minerality dominating the middle palate. Offers compelling sappiness and definition. Noble tannins spread out to saturate the entire palate on the inexorable, steadily building finish.', varietal: 'Cabernet Sauvignon', price: '3,550', size: '750 ml', img:'/images/screaming.jpg', place:{ city: 'Napa Valley', state: 'CA', country: 'USA'}}
-// {name:, vintage:, varietal:, price:, size:, img:, place:{ city:, state:, country:},
+{name: 'Stags Leap', vintage: 2012, quantity: 10, description: 'This Napa Valley Cabernet Sauvignon exemplifies the elegant success that resulted from the cooler 2011 vintage. Black currant, cherry and blueberry aromatics remain distinctly bright in this inviting wine. The palate offers a layered complexity of forest floor, savory herbs, and sweet tobacco that seamlessly blends with the vibrant red fruit. Persistent tannins frame a round, full structure and soft, fresh finish.', varietal: 'Cabernet Saugvinon', price: '107.99', size: '750 ml', img: '/images/stags.png', place: { city: 'Napa Valley', state: 'California', country: 'USA'}},
+{name: 'Screaming Eagle', vintage: 2012, quantity: 4, description: 'Bright medium ruby. Very deep aromas of blueberry, flowers, licorice and wild herbs show a medicinal reserve. Thick but vibrant, with deep, fully ripe, layered flavors of blackberry and blueberry syrup and graphite minerality dominating the middle palate. Offers compelling sappiness and definition. Noble tannins spread out to saturate the entire palate on the inexorable, steadily building finish.', varietal: 'Cabernet Sauvignon', price: '3,550', size: '750 ml', img: '/images/screaming.jpg', place: { city: 'Napa Valley', state: 'CA', country: 'USA'}},
+ {name: 'Trefethen Halo' , vintage: 2008, varietal: 'Cabernet Sauvignon', quantity: 6, description: 'The fruit is incredibly ripe and intense, with flavors of black currants and chocolate-covered cherries. As hard as the tannins are, they must have been even more aggressive two or three years ago, so kudos to Trefethen for holding the bottle back for five years.', price: '219.99', size: '750 ml', img: '/images/halo.jpeg', place:{ city: 'Napa Valley', state: 'California' , country: 'USA'}}, 
 // {name:, vintage:, varietal:, price:, size:, img:, place:{ city:, state:, country:},
 // {name:, vintage:, varietal:, price:, size:, img:, place:{ city:, state:, country:},
 // {name:, vintage:, varietal:, price:, size:, img:, place:{ city:, state:, country:},
